refactor(gui): extract controller + meter wiring from constructor (#3351 Phase 2a)
Seventh PR of the #3351 series and the first of the constructor-
decomposition phase. The MainWindow constructor's "── section ──" wiring
blocks become named private wireXxx() methods defined in their subject
TUs, with the constructor calling each at exactly the original position
— construction order is preserved by definition, so behavior is
identical.

This phase makes the series' first MainWindow.h change: two new private
method declarations (additive only).

- wireExternalControllers() → MainWindow_Controllers.cpp (~590 lines):
  the external-controllers worker thread (#502), FlexControl/MIDI/HID
  manager construction + signal routing, RC-28 deferred press/hold
  (#3323), TMate 2 / Ulanzi / StreamDeck setup.
- wireMeters() → MainWindow_Wiring.cpp (~310 lines): MeterModel →
  S-Meter / Tuner / MTR / HLTH / TX applet routing.

Supporting moves (zero remaining callers in MainWindow.cpp after the
block move; verified by grep):

- tmate2EncoderDefaultAction / tmate2KeyDefaultAction /
  tmate2PushDefaultAction → Controllers TU anon namespace.
- isCwMomentaryActionId → Controllers TU anon namespace (its last
  MainWindow.cpp caller was inside the moved block).

Includes carried explicitly per the Qt 6.4.2-floor lesson from Phase 1e
(IambicKeyer, UlanziDialMapperDialog, the five meter-applet headers,
PeripheralSettings).

MainWindow.cpp: 11,362 → 10,430 lines (series: 19,474 → 10,430, −46.4%).
Constructor: 4,180 → ~3,280 lines.
Verified: full build clean, all 33 ctest suites pass, a11y lint clean.

Refs #3351. Principle XI.
